♻️ refactor: 수정할 부분들 수정함
