+++ services/quiz/src/pages/StudyDetailPage/components/RankingList/style.ts
@@ -1,4 +1,3 @@
-// import { theme } from '@depromeet/theme';
 import styled from '@emotion/styled';
 
 export const List = styled.ul`
